requirements.txt
# ETL Pipeline Requirements
# Python dependencies for the business analysis ETL pipeline

# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# File Processing
openpyxl>=3.0.0
PyPDF2>=3.0.0

# Configuration and Data Formats
PyYAML>=6.0
jsonschema>=4.0.0

# Date and Time Processing
python-dateutil>=2.8.0

# Logging and Utilities
colorlog>=6.0.0

# Testing (optional, for development)
pytest>=7.0.0
pytest-cov>=4.0.0

# Development Tools (optional)
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
